Version 0.9.0 - Let's put this up online
I've been meaning to get this running on a public website for some time.
A lot of the work was done almost a year ago now, and what isn't that
old is at least a few months old by now.

I'm conceding to myself that this is only going to work nicely on
desktops and laptops, and not on mobile without a serious rewrite of.
The big remaining TODO outside of that is the per-world breakdown of
available minerals by type.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@gmail.com>